moved comments check into test package
